Refactored BaseTestMixins into its own package so that it can be used throughout the project
