+++ whoknows/questions/tests/test_views.py
@@ -1,23 +1,8 @@
 from django.test import TestCase
 from django.urls import reverse
-from django.contrib.auth import get_user_model
 from django.contrib import auth
-from questions.models import Question, Tag
-
-
-class BaseTestMixins:
-
-    def create_test_data():
-        User = get_user_model()
-        credentials = {'username': 'john', 'password': 'p@ssw0rd'}
-        user = User.objects.create_user(**credentials)
-        tag1 = Tag.objects.create(name='tag1')
-        tag2 = Tag.objects.create(name='tag2')
-        question1 = Question.objects.create(user=user, title='first question', content='content for first question')
-        question1.tags.add(tag1)
-        question2 = Question.objects.create(user=user, title='second question', content='content for second question')
-        question2.tags.add(tag1)
-        question2.tags.add(tag2)
+from questions.models import Tag
+from tests.mixins import BaseTestMixins
 
 
 class QuestionCreateAndDetailTest(TestCase, BaseTestMixins):
@@ -50,7 +35,7 @@ class QuestionCreateAndDetailTest(TestCase, BaseTestMixins):
         user = auth.get_user(self.client)
         self.assertEqual(user.is_authenticated, False)
         response = self.client.get(reverse('questions:detail', args=['first-question']))
-        # self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.status_code, 200)
 
     def test_invalid_question_detail_displays_404(self):
         response = self.client.get(reverse('questions:detail', args=['this-question-does-not-exist']))
